backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pymysql==1.1.1
groq==0.11.0
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
numpy==1.26.4
python-multipart==0.0.12
httpx==0.27.2
aiofiles==24.1.0
